Moved transaction list to containers folder
