+++ src/app/transactions/component/transaction-detail/transaction-detail.component.ts
@@ -7,7 +7,7 @@ import { Category } from '../../shared/models/category.model';
 import { Transaction } from '../../shared/models/transaction.model';
 import { CategoryService } from '../../shared/services/category.service';
 import { TransactionService } from '../../shared/services/transaction.service';
-import { TransactionListComponent } from '../transaction-list/transaction-list.component';
+import { TransactionListComponent } from '../../container/transaction-list/transaction-list.component';
 
 @Component({
   selector: 'app-transaction-detail',
