+++ src/app/transactions/container/transactions/transactions.component.html
@@ -1,1 +1,1 @@
-<app-transaction-list [data]="transactions"></app-transaction-list>
+<router-outlet></router-outlet>
