+++ src/app/transactions/container/transactions/transactions.component.ts
@@ -1,19 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { Transaction } from '../../shared/models/transaction.model';
-import { TransactionService } from '../../shared/services/transaction.service';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-transactions',
   templateUrl: './transactions.component.html',
   styleUrls: ['./transactions.component.scss']
 })
-export class TransactionsComponent implements OnInit {
-
-  transactions: Transaction[] = [];
-
-  constructor(private transactionService : TransactionService) { }
-
-  ngOnInit(): void {
-    this.transactionService.get().subscribe(response => this.transactions = response);
-  }
-}
+export class TransactionsComponent {}
